requirements.txt
flask
flask-socketio
eventlet
redis
python-socketio[client]
rich
